# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Portfolio API",
        "description": "Laravel12で構築したWeb APIのドキュメントです",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/articles/{article}/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "IDで指定した記事にコメントを投稿する",
                "operationId": "c9bcf038385fccd3ba845a8304296e30",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "記事ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "description": "コメント内容",
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "コメントの内容です。"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "記事が見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "バリデーションエラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required. (and 1 more error)"
                    },
                    "errors": {
                        "description": "各フィールドごとのエラーメッセージ",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The title field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctumで発行されたBearerトークンを入力してください",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}